# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Biblioteca API - Enterprise Edition",
        "description": "Documentación completa y profesional de la API para el sistema de gestión de biblioteca. Incluye módulos de administración avanzada, trazabilidad (Audit Logs), reportes dinámicos y herramientas operativas (QR, Bulk Import).",
        "contact": {
            "email": "soporte@bibliotecaapi.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de API Principal"
        }
    ],
    "paths": {
...
        "/api/loans": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "List all loans (admin) or user's own loans",
                "operationId": "53fdef403307586122900dc3b1d07368",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Loan"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Solicitar un préstamo (usuario autenticado)",
                "operationId": "92bd5c653a9a8345b4f161d5ba3bbd71",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "book_id"
                                ],
                                "properties": {
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Préstamo aprobado automáticamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Préstamo aprobado automáticamente."
                                        },
                                        "loan": {
                                            "$ref": "#/components/schemas/Loan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No puedes solicitar el préstamo"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/loans/{loan}": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Get a specific loan",
                "operationId": "39a4e10f305b5be4c3e4eba31a94777a",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Loan found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Loans"
                ],
                "summary": "Actualizar préstamo (solo admin). Si se establece return_date por primera vez, se devuelve el libro y se aumenta el stock.",
                "operationId": "d1a9dfd54e508debb4769bc440e372f8",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 11
                                    },
                                    "loan_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-01",
                                        "nullable": true
                                    },
                                    "return_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-15",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "aprobado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Préstamo actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Loan"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Loans"
                ],
                "summary": "Delete loan (admin only)",
                "operationId": "6d80bd9e85d71ce0f714269e753f85fb",
                "parameters": [
                    {
                        "name": "loan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Loan deleted"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/loans/{id}/return": {
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Marcar préstamo como devuelto (solo admin)",
                "operationId": "0531830e393b1b9b492e67a2106b7b31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del préstamo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Préstamo marcado como devuelto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Préstamo marcado como devuelto."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Préstamo no encontrado o ya devuelto"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Loan": {
                "required": [
                    "id",
                    "user_id",
                    "book_id",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "book_id": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "aprobado",
                            "devuelto"
                        ]
                    },
                    "loan_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "return_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}